Add support for nameservers with IPv6 addresses
This change fixes a couple of places in the kube-dns code where
IPv6 addresses are not being handled correctly for nameservers found
in a minion's resolv.conf file.

This change is needed in order to get kube-dns to operate in a
Kubernetes cluster that is IPv6-only. See Kubernetes issue 51371
(https://github.com/kubernetes/kubernetes/issues/51371).

fixes #147