Build and push manifests for container images
We build and push images for various architectures, but are not
publishing a single manifest that can be used to look up the images.
test/images in main k/k repository and kube images in k/release already
publish manifests for all the image we need for e2e testing, so we
should do the same as the 2 images in this repository are part of the
conformance e2e test suite.