support kube-dns stubdomain config in node-cache
This change adds functionality in node-cache to generate
the Corefile based on stubDomain and upstream servers config
in kube-dns config map. This requires the kube-dns configmap
to be mounted in the local filesystem.

This change also supports using a custom service for clusterDNS
This service name is kube-dns by default.