move to go modules, use coreDNS 1.7, ginkgo1.14.10
Also set GOCACHE, but disable caching for builds.

Setting GOCACHE is required in go 1.12 onwards. But the cache dirs
are created as root and run into permission issues if build is run
by a non-root user. Build cache should also be disabled for CI, so
this change sets GOCACHE but does not use it in docker run.